Cast to uint8 before taking argmax

+++ attacks/fgsm_I3_12xMN1_multitop_logits_normalise_063_sneakycondprob/attack_fgsm.py
@@ -356,7 +356,7 @@ def main(_):
         class_is_next_highest = tf.logical_and(
             tf.concat([vec_true, class_is_before_threshold[:, :-1]], -1),
             tf.logical_not(class_is_before_threshold))
-        index_next_highest = tf.argmax(class_is_next_highest, axis=-1)
+        index_next_highest = tf.argmax(tf.cast(class_is_next_highest, tf.uint8), axis=-1)
         index_very_highest = sort_indices[:, 0]
 
         target_indices = tf.stack((index_very_highest, index_next_highest), axis=-1)
